Added a line to alter the max_map_count limit.
Elasticsearch requires vm.max_map_count set to min of 262144, otherwise it
fails to start in kubernetes cluster. I've tried to preset the value with
init-containers which do not seem to work for replication controllers (they're
still in beta accessible only through annotations).

Change here just sets these values to the min. requires rebuilding both
docker-elasticsearch and docker-elasticsearch-kubernetes which uses this
image as a base.